Resolving test
